added FITS and HDU maps

+++ mojofitsio/byteio/general.mojo
@@ -3,19 +3,19 @@ from math.bit import bswap
 
 from mojofitsio.config import config
 
-fn get_FITS_blocks(file: FileHandle, startblock: Int) raises -> Tensor[DType.int8]:
+fn get_FITS_blocks(file: FileHandle, startblock: UInt64) raises -> Tensor[DType.int8]:
     """
     Return the FITS block at an offset of *startblock* FITS blocks.
     """
     var bc = file.seek(config.fits_block_length*startblock)
     return file.read_bytes(config.fits_block_length)
 
-fn get_FITS_blocks(file: FileHandle, startblock: Int, endblock: Int) raises -> Tensor[DType.int8]:
+fn get_FITS_blocks(file: FileHandle, startblock: UInt64, endblock: UInt64) raises -> Tensor[DType.int8]:
     """
     Return an integral number of FITS blocks in the range [*startblock*, *endblock*).
     """
     var bc = file.seek(config.fits_block_length*startblock)
-    return file.read_bytes(config.fits_block_length*(endblock - startblock))
+    return file.read_bytes(config.fits_block_length*int((endblock - startblock)))
 
 fn i16_frombytes[inwidth: Int, outwidth: Int = inwidth//2](data: SIMD[DType.int8, inwidth]) -> SIMD[DType.int16, outwidth]:
     """
